Add a simple in-memory data store.
There are two global state objects, both sets. The first is simply a
set of PhoneNumberEntry records, meant to represent a simple database
table. The second is a set of PhoneNumberUniqueIndexEntry objects,
meant to represent a unique index on the table.

The add-entry function will throw an exception when a caller tries to
add a duplicate entry.