Refactor state and fix a bug.
Move all the state into one reference to avoid any possibility of race
conditions when checking the index.

Move some logic into separate functions for better readibility.

With clarification that first entry should win when there are
duplicates in seed data, finalize init logic.

Fix a bug in ensure-entry (map to PhoneNumberEntry was wrong because
sequence was created from the map in the wrong order).